refactor: rename files - Ref gestion-de-projet#2027
