+++ src/components/NewScopeTree/CohortCreationCareSite/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
-const CohortCreationCareSite = (props: any) => {
+const Index = (props: any) => {
   return <div>Composant permettant la selection d'une population source ou bien d'une unite executive{props}</div>
 }
 
-export default CohortCreationCareSite
+export default Index
