+++ src/components/NewScopeTree/commons/styles.js
