+++ src/views/Scope/CareSiteView.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react'
-import CareSiteExploration from 'components/NewScopeTree/ExploratedCareSite/CareSiteExploration'
+import CareSiteExploration from 'components/NewScopeTree/CareSiteExploration'
 import { Button, Grid } from '@mui/material'
 import ScopeSearchBar from 'components/Inputs/ScopeSearchBar/ScopeSearchBar'
-import CareSiteSearch from 'components/NewScopeTree/CareSiteSearch/CareSiteSearch'
+import CareSiteSearch from 'components/NewScopeTree/CareSiteSearch/index'
 import useStyles from './styles'
 import { useNavigate } from 'react-router-dom'
 import { useAppDispatch, useAppSelector } from '../../state'
